Fix test fixture

+++ test/Typesense.Tests/TypesenseFixture.cs
@@ -2,19 +2,15 @@ using Microsoft.Extensions.DependencyInjection;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Typesense.Setup;
+using Xunit;
 
 namespace Typesense.Tests;
 
-public class TypesenseFixture
+public class TypesenseFixture : IAsyncLifetime
 {
     public ITypesenseClient Client => GetClient();
 
-    public TypesenseFixture()
-    {
-        Cleanup().ContinueWith(_ => { });
-    }
-
-    private async Task Cleanup()
+    public async Task InitializeAsync()
     {
         await CleanCollections();
         await CleanApiKeys();
@@ -43,6 +39,11 @@ public class TypesenseFixture
     private async Task CleanCurations()
     {
         var curations = await Client.ListSearchOverrides("companies");
+        if (curations == null)
+        {
+            return;
+        }
+
         foreach (var curation in curations.SearchOverrides)
         {
             await Client.DeleteSearchOverride("companies", curation.Id);
@@ -74,5 +75,10 @@ public class TypesenseFixture
                     }
                 };
             }).BuildServiceProvider().GetService<ITypesenseClient>();
+    }
+
+    public Task DisposeAsync()
+    {
+        return Task.CompletedTask;
     }
 }
